chore(clerk-js,shared,nextjs): Move `setDevBrowserJWTInURL` to shared
