+++ packages/shared/.gitignore
@@ -4,6 +4,7 @@ color
 cookie
 date
 deprecated
+devBrowser
 error
 file
 globs
