+++ packages/shared/package.json
@@ -51,6 +51,7 @@
     "cookie",
     "date",
     "deprecated",
+    "devBrowser",
     "error",
     "file",
     "globs",
