+++ packages/shared/src/devBrowser.ts
@@ -1,12 +1,4 @@
-// TODO: This is a partial duplicate of part of packages/clerk-js/src/utils/devBrowser.ts
-// TODO: To be removed when we can extract this utility to @clerk/shared
-
 export const DEV_BROWSER_SSO_JWT_PARAMETER = '__dev_session';
-
-//
-// Below this line should be identical to clerk-js version
-//
-
 export const DEV_BROWSER_JWT_MARKER = '__clerk_db_jwt';
 const DEV_BROWSER_JWT_MARKER_REGEXP = /__clerk_db_jwt\[(.*)\]/;
 
@@ -14,17 +6,8 @@ const DEV_BROWSER_JWT_MARKER_REGEXP = /__clerk_db_jwt\[(.*)\]/;
 export function setDevBrowserJWTInURL(url: URL, jwt: string, asQueryParam: boolean): URL {
   const resultURL = new URL(url);
 
-  // extract & strip existing jwt from hash
-  const jwtFromHash = extractDevBrowserJWTFromHash(resultURL.hash);
-  resultURL.hash = resultURL.hash.replace(DEV_BROWSER_JWT_MARKER_REGEXP, '');
-  if (resultURL.href.endsWith('#')) {
-    resultURL.hash = '';
-  }
-
-  // extract & strip existing jwt from search
-  const jwtFromSearch = resultURL.searchParams.get(DEV_BROWSER_SSO_JWT_PARAMETER);
-  resultURL.searchParams.delete(DEV_BROWSER_SSO_JWT_PARAMETER);
-
+  const jwtFromHash = extractDevBrowserJWTFromURLHash(resultURL);
+  const jwtFromSearch = extractDevBrowserJWTFromURLSearchParams(resultURL);
   // Existing jwt takes precedence
   const jwtToSet = jwtFromHash || jwtFromSearch || jwt;
 
@@ -42,4 +25,29 @@ export function setDevBrowserJWTInURL(url: URL, jwt: string, asQueryParam: boole
 function extractDevBrowserJWTFromHash(hash: string): string {
   const matches = hash.match(DEV_BROWSER_JWT_MARKER_REGEXP);
   return matches ? matches[1] : '';
+}
+
+/**
+ * Extract & strip existing jwt from hash
+ * Side effect: Removes dev browser from the url hash
+ **/
+export function extractDevBrowserJWTFromURLHash(url: URL) {
+  const jwt = extractDevBrowserJWTFromHash(url.hash);
+  url.hash = url.hash.replace(DEV_BROWSER_JWT_MARKER_REGEXP, '');
+  if (url.href.endsWith('#')) {
+    url.hash = '';
+  }
+
+  return jwt;
+}
+
+/**
+ * Extract & strip existing jwt from search params
+ * Side effect: Removes dev browser from the search params
+ **/
+export function extractDevBrowserJWTFromURLSearchParams(url: URL) {
+  const jwtFromDevSession = url.searchParams.get(DEV_BROWSER_SSO_JWT_PARAMETER);
+  url.searchParams.delete(DEV_BROWSER_SSO_JWT_PARAMETER);
+
+  return jwtFromDevSession || '';
 }
